+++ CHANGES.md
@@ -6,6 +6,7 @@
   * Add `group_summary` transformers.
 * Fix warning due to changing default `dtype` when creating an empty
   `pandas.Series`.
+* Enable `GridSearch` for more than two sensitive features values.
 
 ### v0.4.5
 * Changes to `ThresholdOptimizer`: